ngrx cleaning up leftovers
