+++ src/app/landing/landing.component.html
@@ -12,12 +12,6 @@
 
     <mat-sidenav-content class="d-flex justify-content-center align-items-center">
       <h2 class="text-secondary">Stay tuned...</h2>
-      <button aria-labelledby="signIn" mat-flat-button color="accent" [routerLink]="['restrict']"
-        routerLinkActive="router-link-active">
-        <span id="signIn">Sign in</span>
-      </button>
-      <pre>
-      </pre>
       <router-outlet></router-outlet>
     </mat-sidenav-content>
